Removes MonadThrow
I've changed my mind about this. I want the functions to be total
and return Either DockerError a. After some thinking, I don't think
the ExceptT IO antipattern pertains to this case becuase ExceptT isn't
exposed to the user rather the user get's Either Error a. The only one
who'll see the ExceptT are the libray maintainers.

The idea is to catch all the relevant exceptions in the defaultHttpHandler
so that the user doesn't have to care that they are dealing with a HTTP Api
because that's totally an implementation detail.

I realize that I'm sort of going the catch-all/mask-all road here but
I will only handle the most common cases to abstract away that we're
talking to Docker over a remote http api. I won't mask all, ie. I won't
catch SomeException and therefore break async exceptions that way.
For 99% of the cases this should be fine but for more advanced users
there's always the option of building their own httpHandler to handle
whatever special use-case they might have and the functions won't care
because they are polymorphic Monad wise.