Remove to/from json instances for Volume
We only provide instances for a list of Volumes
and the interface for adding those to createOpts
is via the addVolume function. Everything else will
be caught at compile time whereas the fromJson instance for
a single Volume was flawed (if nothing then because it was
using head and not checking for an empty list,
or enforcing only one element in the list).