Fix bug where an Image's RepoTags is null
I'm not sure if this is perhaps a docker bug because it seemed that my
docker daemon got perhaps in an illegal state. My assumption so far was
always that if repo tags was null that the image will assume the tag
"latest". Be as it may we have to support this case for now otherwise
decoding the image list blows up.