Bumps C in version string
According to the PVP:

Backwards compatible client specification. Often a package maintainer wants
to add to an API without breaking backwards compatibility, and in that case
they can follow the rules of point 2, and increase only C. A client MAY specify
that they are insensitive to additions to the API by allowing a range of C
values, e.g. build-depends: base >= 2.1.1 && < 2.2.